Added documantation, corrected tools.rst, minor change in hologram_image

+++ hyperspy/_signals/hologram_image.py
@@ -224,14 +224,14 @@ class HologramImage(Signal2D):
         # TODO: Add smoothing of Fresnel filter
         # TODO: Add other smoothing options?
 
-        image_size = sb_size*2
+        image_size = np.int(sb_size*2)
         (sx, sy) = holo_data.shape
         holo_data = np.float64(holo_data)
 
         h_hw_fft = fftshift(fft2(holo_data))  # <---- NO Hanning filtering
 
-        sb_roi = h_hw_fft[np.int(sb_pos[0]-image_size//2):np.int(sb_pos[0]+image_size//2),
-                 np.int(sb_pos[1]-image_size//2):np.int(sb_pos[1]+image_size//2)]
+        sb_roi = h_hw_fft[(sb_pos[0]-image_size//2):(sb_pos[0]+image_size//2),
+                 (sb_pos[1]-image_size//2):(sb_pos[1]+image_size//2)]
 
         (sb_ny, sb_nx) = sb_roi.shape
         sb_l = min(sb_ny / 2, sb_nx / 2)
@@ -267,10 +267,10 @@ class HologramImage(Signal2D):
         sinc_k = 5.0  # Sink times SBsize
         w_one = np.sinc(
             np.linspace(-sb_ny/2, sb_ny/2, sb_ny) * np.pi / (sinc_k * sb_size))
-        w_one = w_one.reshape((np.int(image_size), 1))
+        w_one = w_one.reshape(image_size, 1)
         w_two = np.sinc(
             np.linspace(-sb_nx/2, sb_nx/2, sb_nx) * np.pi / (sinc_k * sb_size))
-        window = w_one.dot(w_two.reshape((1, np.int(image_size))))
+        window = w_one.dot(w_two.reshape(1, image_size))
 
         # IFFT
         wav = ifft2(fftshift(sb_roi * c_mask * np.logical_not(f_mask) * window))
